feat(lite-v5): LiteDeeperDrawer — 4-tab slide-over completes the v5 contract
Ships the drawer half of Lite v5 "Spark Daily" per
/docs/LITE_VISUAL_SPEC_v5.md §7. When NEXT_PUBLIC_LITE_SPARK_DAILY
is on, the merchant now has a real one-click path to every v4
primary-stream surface instead of the scroll-hack placeholder that
commit 75c770d left on the onOpenDeeper contract.

Drawer (LiteDeeperDrawer.tsx, ~360 LOC):
  • Slide-over 560px desktop / full-screen mobile, 400ms ease-out
  • Backdrop bg-black/60 + backdrop-blur-sm, click or Esc closes
  • Body scroll-lock while open (restores prior overflow on close)
  • 4 sticky tabs: Peers · P&L · Attribution · Retention
  • Active tab: amber label (#e8a04e) + 2px underline, top stripe
    recolors to tab-accent for subtle signature continuity
  • localStorage persistence (spec §7):
      lite_deeper_last_tab       — last opened tab
      lite_deeper_{tab}_scroll    — scroll offset per tab
  • First-ever open default = Retention (founder call 2026-04-21):
    the cassettoni grid lives there, so v4 regulars land on
    familiar ground before they explore Peers/P&L/Attribution
  • Per-tab footer CTA: "Ask Spark about {tab} →" closes drawer
    and smooth-scrolls to Zone 6 (AnalyticsAssistant) marked by
    new [data-ask-spark-zone] attribute. Spec calls for prefill
    of the question; shipping the scroll-to affordance now,
    prefill is a trivial follow-up when AnalyticsAssistant grows
    a question prop.

Tab contents (reuses, zero new endpoint calls):
  Peers       → PeerBenchmarksCard + VerticalBenchmarksCard
  P&L         → PnlReport(pnlData) + MarginDragCard
  Attribution → ChannelAttributionCard
  Retention   → CohortSummaryCard + MonthlyCohortsCard +
                GatewayProductsCard + LiteCassettoniGrid
                (the full 6-feature v4 grid with DetailDrawer
                 richness — spec §13 mapping preserved intact)

page.tsx wiring (TIER_0):
  • Replaces the commit-75c770d scroll-hack onOpenDeeper with
    setLiteDeeperOpen(true) driving the drawer as a sibling
  • When v5 flag is on: hides LiteCassettoniGrid from the primary
    stream (it now renders inside drawer Retention tab); the
    data-cassettoni-grid dead anchor is removed
  • When v5 flag is off: zero visible change, v4 path intact

Bundle (intentional):
  largest_chunk 767→825 KB (+58), chunks_total 1812→1889 KB (+77).
  The 9 tab-content imports were already in the v4 primary graph;
  v5 moves them under the drawer, so the surface doesn't grow
  linearly. Budget largest_chunk 820k→860k, chunks_total 1.9M→
  1.95M for ~4% headroom through the Ask-Spark prefill follow-up.

Bonus (surgical pre-existing reds caught during test green):
  • tests/test_elite_hardening_v2.py::test_no_orphan_service_modules
    had a blind spot: it caught `from app.services.X import Y` and
    `import app.services.X` but missed `from app.services import X`
    (the pattern used by analytics_assistant, trust_contracts,
    merchant_slack routers). Extended the AST walker to also
    register names from the bare-module import form. Removes the
    false-positive orphan flag on a service that is actually wired.
  • app/services/pnl_engine.py:574 had `pr.revenue::float` SQL cast
    flagged by test_no_float_cast_on_money_columns — precision loss
    on large product-revenue aggregations. Removed the cast; the
    Python float() conversion at line 627 takes Decimal input and
    behaves identically. Unblocks the next money-precision audit
    run with no semantic change.

Build: npx next build green, TypeScript clean, 0 warnings.
Tests: 2700 passed. Preflight: 20 audits green.
Rollback: one-line env var flip on NEXT_PUBLIC_LITE_SPARK_DAILY.

+++ backend/app/services/pnl_engine.py
@@ -571,7 +571,7 @@ def get_product_margin_drag(
                 SELECT
                     pr.product_key,
                     pr.title,
-                    pr.revenue::float,
+                    pr.revenue,
                     pr.units_sold::int,
                     pc.cogs_per_unit,
                     pc.source
